+++ docs/examples/auto-instrumentation/server_manual.py
@@ -14,7 +14,9 @@
 
 from flask import Flask, request
 
-from opentelemetry import trace
+from opentelemetry.trace import (
+    set_tracer_provider, get_tracer_provider, SpanKind
+)
 from opentelemetry.instrumentation.wsgi import collect_request_attributes
 from opentelemetry.propagate import extract
 from opentelemetry.sdk.trace import TracerProvider
@@ -25,10 +27,10 @@ from opentelemetry.sdk.trace.export import (
 
 app = Flask(__name__)
 
-trace.set_tracer_provider(TracerProvider())
-tracer = trace.get_tracer_provider().get_tracer(__name__)
+set_tracer_provider(TracerProvider())
+tracer = get_tracer_provider().get_tracer(__name__)
 
-trace.get_tracer_provider().add_span_processor(
+get_tracer_provider().add_span_processor(
     BatchSpanProcessor(ConsoleSpanExporter())
 )
 
@@ -38,7 +40,7 @@ def server_request():
     with tracer.start_as_current_span(
         "server_request",
         context=extract(request.headers),
-        kind=trace.SpanKind.SERVER,
+        kind=SpanKind.SERVER,
         attributes=collect_request_attributes(request.environ),
     ):
         print(request.args.get("param"))